Split apart actions and goals
